// src/registry/lib/datasource-types.ts
export interface TimeRangeConfig {
  type?: 'forward' | 'backward' | 'custom'
  range?: {
    radioType?: number
    gte?: string
    lte?: string
    date?: string
    financialMonth?: {
      date?: number
      time?: string
    }
    record?: {
      table?: { id?: string }
      id?: string
    }
  }
  unit?: string
  count?: number
  fromNow?: boolean
  historyTime?: any
}

/**
 * 时间线配置（用于实时数据）
 */
export interface TimeLineConfig {
  count?: number
  unit?: string
}

/**
 * 分组配置
 */
export interface GroupConfig {
  count?: number
  unit?: string
  fill?: { value: string }
  isClassMode?: boolean
  table?: any
  record?: any
}